Refactor to use HOC for data fetching
